chore: refactor automock utility
